backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
mongoengine
pymongo
dotenv